DCTN-951: added improvements

+++ src/Service/ReportService.php
@@ -19,59 +19,36 @@ use Symfony\Component\Security\Core\User\UserInterface;
 
 class ReportService extends BaseService
 {
-    protected FormFactoryInterface $formFactory;
     protected ReportRepository $reportRepository;
     protected FileUploader $fileUploader;
     protected ConferenceService $conferenceService;
     protected EntityManagerInterface $entityManager;
     protected FlashBagInterface $flashBag;
+    protected Filesystem $filesystem;
 
     public function __construct(
-        FormFactoryInterface   $formFactory,
         ReportRepository       $reportRepository,
         FileUploader           $fileUploader,
         ConferenceService      $conferenceService,
         EntityManagerInterface $entityManager,
-        FlashBagInterface      $flashBag
+        FlashBagInterface      $flashBag,
+        Filesystem             $filesystem
     )
     {
-        $this->formFactory = $formFactory;
         $this->reportRepository = $reportRepository;
         $this->fileUploader = $fileUploader;
         $this->conferenceService = $conferenceService;
         $this->entityManager = $entityManager;
         $this->flashBag = $flashBag;
+        $this->filesystem = $filesystem;
     }
 
-    /**
-     * @return array<Report>
-     */
-    public function getAllReports(): array
-    {
-        return $this->reportRepository->findAll();
-    }
-
-    public function prepareForm(
-        Report     $report,
-        Request    $request,
-        Conference $conference
-    ): FormInterface
-    {
-        $form = $this->formFactory->create(ReportType::class, $report, [
-            'conference_id' => $conference->getId(),
-            'conference_start' => $conference->getStartedAt(),
-            'conference_end' => $conference->getEndedAt()
-        ]);
-        $form->handleRequest($request);
-
-        return $form;
-    }
-
-    public function saveReport(
+    public function save(
         Report        $report,
         Conference    $conference,
         UserInterface $user,
-        ?UploadedFile $document = null
+        ?UploadedFile $document = null,
+        bool          $api = false
     ): ?Report
     {
         if ($document) {
@@ -79,56 +56,24 @@ class ReportService extends BaseService
 
             if ($reportId) {
                 $fileExist = $this->reportRepository->fileNameExist($reportId);
-                $this->deleteUploadedFile(array_shift($fileExist));
+
+                if ($fileExist) {
+                    $this->deleteUploadedFile(array_shift($fileExist));
+                }
             }
 
             $documentName = $this->fileUploader->upload($document);
 
             if (!$documentName) {
+                if ($api) {
+                    throw new \RuntimeException("File upload failed due to server error.", 500);
+                }
                 $this->flashBag->add(
                     'file-error',
                     'File upload error. Try again later.'
                 );
 
                 return null;
-            }
-
-            $report->setDocument($documentName);
-        }
-
-        $userPartOfConference = $this->conferenceService->findParticipantByUserId($user->getId(), $conference->getId());
-
-        if (!$userPartOfConference) {
-            $conference->addUser($user);
-        }
-
-        $report->setUser($user);
-        $report->setConference($conference);
-        $this->reportRepository->saveData($report);
-
-        return $report;
-    }
-
-    public function saveReportApi(
-        Report        $report,
-        Conference    $conference,
-        UserInterface $user,
-        ?UploadedFile $document = null
-    ): Report
-    {
-        if ($document) {
-
-            $reportId = $this->entityManager->contains($report) ? $report->getId() : null;
-
-            if ($reportId) {
-                $fileExist = $this->reportRepository->fileNameExist($reportId);
-                $this->deleteUploadedFile(array_shift($fileExist));
-            }
-
-            $documentName = $this->fileUploader->upload($document);
-
-            if (!$documentName) {
-                throw new \RuntimeException("File upload failed due to server error.", 500);
             }
 
             $report->setDocument($documentName);
@@ -154,11 +99,10 @@ class ReportService extends BaseService
 
     public function deleteUploadedFile(string $fileName): void
     {
-        $filesystem = new Filesystem();
         $filePath = $this->fileUploader->getTargetDirectory() . '/' . $fileName;
 
-        if ($filesystem->exists($filePath)) {
-            $filesystem->remove($filePath);
+        if ($this->filesystem->exists($filePath)) {
+            $this->filesystem->remove($filePath);
         }
     }
 
@@ -171,17 +115,23 @@ class ReportService extends BaseService
     {
         $filePath = $this->fileUploader->getTargetDirectory() . '/' . $fileName;
 
-        if (!file_exists($filePath) || !is_readable($filePath)) {
-            throw new NotFoundHttpException('File not found or unavailable');
+        if (!file_exists($filePath)) {
+            throw new NotFoundHttpException('File not found');
+        } else {
+            $filePerms = fileperms($filePath);
+
+            if (!(($filePerms & 0x0100) === 0x0100)
+                ||
+                !(($filePerms & 0x0020) === 0x0020)
+                ||
+                !(($filePerms & 0x0004) === 0x0004)) {
+                throw new NotFoundHttpException('File unavailable');
+            }
         }
 
         $response = new StreamedResponse(function () use ($filePath) {
             $stream = fopen($filePath, 'r');
 
-            if ($stream === false) {
-                throw new \RuntimeException('Failed to open the file for reading');
-            }
-
             while (!feof($stream)) {
                 echo fread($stream, 1024);
                 flush();
@@ -190,32 +140,6 @@ class ReportService extends BaseService
             fclose($stream);
         });
 
-        $response->headers->set('Content-Type', 'application/octet-stream');
-        $response->headers->set('Content-Disposition', "attachment; filename=\"{$fileName}\"");
-
-        return $response;
-    }
-
-    public function downloadFileApi(string $fileName): StreamedResponse
-    {
-        $response = new StreamedResponse(function () use ($fileName) {
-            $file = $this->fileUploader->getTargetDirectory() . '/' . $fileName;
-
-            if (!file_exists($file)) {
-                throw $this->createNotFoundException('File not found.');
-            }
-
-            $stream = fopen($file, 'r');
-
-            while (!feof($stream)) {
-                echo fread($stream, 1024);
-                flush();
-            }
-
-            fclose($stream);
-        });
-
-        // Встановлення заголовків для завантаження файлу
         $response->headers->set('Content-Type', 'application/octet-stream');
         $response->headers->set('Content-Disposition', "attachment; filename=\"{$fileName}\"");
 
@@ -229,14 +153,18 @@ class ReportService extends BaseService
 
     public function prepareReportFilters(array $filters, Conference $conference): array
     {
-        if ($filters['start_time']->format('Y-m-d H:i:s') === $filters['end_time']->format('Y-m-d H:i:s')) {
+        if (
+            (!empty($filters['start_time']) && !empty($filters['end_time']))
+            &&
+            ($filters['start_time']->format('Y-m-d H:i:s') === $filters['end_time']->format('Y-m-d H:i:s'))
+        ) {
             $filters['start_time'] = null;
             $filters['end_time'] = null;
 
             return $filters;
         }
 
-        if ($filters['start_time']) {
+        if (!empty($filters['start_time'])) {
             $filters['start_time'] = (new \DateTime($conference->getStartedAt()->format('Y-m-d')))->setTime(
                 $filters['start_time']->format('H'),
                 $filters['start_time']->format('i'),
@@ -244,7 +172,7 @@ class ReportService extends BaseService
             );
         }
 
-        if ($filters['end_time']) {
+        if (!empty($filters['end_time'])) {
             $filters['end_time'] = (new \DateTime($conference->getEndedAt()->format('Y-m-d')))->setTime(
                 $filters['end_time']->format('H'),
                 $filters['end_time']->format('i'),
